child-node/requirements.txt
torch
torchvision
transformers
accelerate
fastapi
uvicorn
Pillow
safetensors
